requirements.txt
streamlit>=1.35.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
PyPDF2>=3.0.1
numpy>=1.26.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0